requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
apturl==0.5.2
argcomplete==3.1.4
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.4.0
autopep8==2.0.4
bcc==0.29.1
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.0
Brlapi==0.8.5
bs4==0.0.2
cachetools==6.2.2
catalogue==2.0.10
certifi==2023.11.17
cffi==2.0.0
chardet==5.2.0
click==8.1.6
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
command-not-found==0.3
confection==0.1.5
configobj==5.0.8
contourpy==1.3.2
cryptography==41.0.7
cupshelpers==1.0
customtkinter==5.2.2
cycler==0.12.1
cymem==2.0.11
darkdetect==0.8.0
dbus-python==1.3.2
decorator==5.2.1
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
dnspython==2.8.0
executing==2.2.1
Flask==3.1.1
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.59.0
freetype-py==2.4.0
Frozen-Flask==1.0.2
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.52.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
gyp==0.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
idna==3.6
importlib-metadata==4.12.0
indicator-cpufreq==0.2.2
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jaraco.classes==3.2.1
jax==0.7.1
jaxlib==0.7.1
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.6
jiter==0.12.0
jupyterlab_widgets==3.0.15
keyring==24.3.1
keyrings.alt==5.0.0
kiwisolver==1.4.8
langcodes==3.5.0
language-selector==0.1
language_data==1.3.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libevdev==0.5
louis==3.29.0
Mako==1.3.2.dev0
marisa-trie==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mate-menu==22.4.2
mate-tweak==22.10.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
meson==1.3.2
ml_dtypes==0.5.3
more-itertools==10.2.0
MouseInfo==0.1.3
murmurhash==1.0.13
mysql-connector-python==8.0.15
netaddr==0.8.0
networkx==3.5
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
onboard==1.4.1
openai==2.8.1
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==24.0
pandas==2.3.3
parso==0.8.5
pep8==1.7.1
pexpect==4.9.0
phonenumbers==9.0.17
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
preshed==3.0.10
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycairo==1.25.1
pycodestyle==2.11.1
pycparser==2.23
pycryptodomex==3.20.0
pycups==2.0.1
pydantic==2.12.4
pydantic_core==2.41.5
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.9.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyqtgraph==0.13.7
PyRect==0.2.0
PyScreeze==1.0.1
python-apt==2.7.7+ubuntu5.1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.2.1
python-whois==0.9.6
python-xlib==0.33
python3-xlib==0.15
pytube==15.0.0
pytweening==1.2.0
pytz==2025.2
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
requests-unixsocket==0.3.0
rich==13.7.1
rlPyCairo==0.3.0
rsa==4.9.1
scipy==1.16.3
screen_brightness_control==0.24.3
SecretStorage==3.3.3
sentencepiece==0.2.1
setproctitle==1.3.3
setuptools==68.1.2
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smart_open==7.4.4
sniffio==1.3.1
sounddevice==0.5.3
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.1
ssh-import-id==5.11
stack-data==0.6.3
systemd-python==235
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.7
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
vboxapi==1.0
wadllib==1.3.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.2
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.42.0
widgetsnbextension==4.0.14
wrapt==2.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.1
